Replaced .env with .env.sample,
added comments on how to load environment variables from .env. The line
`require('dotenv').config();`
with app.js will load the dotenv module. The .env file is under
gitignore, so the users has to make a copy of the .env.sample
file and rename it to .env so that they can use their own postgres
username and password without commiting it into version control.